fix: remove si_prefix.tests import test
As part of version [1.3.3][1], the `si_prefix.tests` module was refactored out into a
top-level repo directory to better follow best practices. However, this change caused
the respective import test to fail in this feedstock Conda recipe.

In this commit, remove the `si_prefix.tests` import test to align with the new 1.3.3
project structure.

[1]: https://github.com/cfobel/si-prefix/releases/tag/v1.3.3